chore: refactor layer manager location
