+++ src/components/Popup/Popup.tsx
@@ -7,9 +7,9 @@ import {Portal} from '../Portal';
 import type {DOMProps, QAProps} from '../types';
 import {FocusTrap, useParentFocusTrap} from '../utils/FocusTrap';
 import {block} from '../utils/cn';
+import {useLayer} from '../utils/layerManager';
+import type {LayerExtendableProps} from '../utils/layerManager';
 import {getCSSTransitionClassNames} from '../utils/transition';
-import {useLayer} from '../utils/useLayer';
-import type {LayerExtendableProps} from '../utils/useLayer';
 import {usePopper} from '../utils/usePopper';
 import type {PopperAnchorRef, PopperPlacement, PopperProps} from '../utils/usePopper';
 import {useRestoreFocus} from '../utils/useRestoreFocus';
