+++ src/components/index.ts
@@ -51,9 +51,8 @@ export * from './utils/class-transform';
 export * from './utils/eventBroker';
 export {getComponentName} from './utils/getComponentName';
 export * from './utils/withEventBrokerDomHandlers';
-export * from './utils/useLayer';
+export * from './utils/layerManager';
 export {Lang, configure} from './utils/configure';
 export {useOnFocusOutside} from './utils/useOnFocusOutside';
 export * from './utils/interactions';
 export * from './utils/xpath';
-export {getLayersCount} from './utils/LayerManager';
