+++ src/components/utils/layerManager/LayerManager.ts
@@ -1,8 +1,7 @@
 import type {VirtualElement} from '@popperjs/core';
 
-import {KeyCode} from '../../constants';
-
-import {eventBroker} from './eventBroker';
+import {KeyCode} from '../../../constants';
+import {eventBroker} from '../eventBroker';
 
 export type LayerCloseReason = 'outsideClick' | 'escapeKeyDown';
