+++ src/components/utils/layerManager/index.ts
@@ -1,0 +1,9 @@
+export {layerManager, getLayersCount} from './LayerManager';
+export type {
+    LayerConfig,
+    LayerExtendableProps,
+    ContentElement,
+    LayerCloseReason,
+} from './LayerManager';
+export {useLayer} from './useLayer';
+export type {LayerProps} from './useLayer';
